+++ examples/refsi/refsi_g1_wi/compiler/refsi_g1_wi/source/refsi_pass_machinery.cpp
@@ -130,8 +130,8 @@ llvm::ModulePassManager RefSiG1PassMachinery::getLateTargetPasses() {
   // addLateBuiltinsPasses, which isn't ideal.
   PM.addPass(compiler::utils::DefineMuxDmaPass());
 
-  // We don't run the barrier pass, make sure that's taken into account.
-  tuner.handling_barriers = false;
+  // We don't run the WorkItemLoopsPass; make sure that's taken into account.
+  tuner.handling_work_item_loops = false;
   addPreVeczPasses(PM, tuner);
 
   addLateBuiltinsPasses(PM, tuner);