+++ modules/compiler/cookie/{{cookiecutter.target_name}}/source/{{cookiecutter.target_name}}_pass_machinery.cpp
@@ -25,7 +25,6 @@
 #include <compiler/utils/align_module_structs_pass.h>
 #include <compiler/utils/attributes.h>
 #include <compiler/utils/encode_kernel_metadata_pass.h>
-#include <compiler/utils/handle_barriers_pass.h>
 #include <compiler/utils/link_builtins_pass.h>
 #include <compiler/utils/metadata.h>
 #include <compiler/utils/metadata_analysis.h>
@@ -35,6 +34,7 @@
 {% endif -%}
 #include <compiler/utils/simple_callback_pass.h>
 #include <compiler/utils/verify_reqd_sub_group_size_pass.h>
+#include <compiler/utils/work_item_loops_pass.h>
 #include <llvm/ADT/StringRef.h>
 #include <llvm/Passes/PassBuilder.h>
 #include <llvm/Transforms/Utils/Cloning.h>
@@ -196,10 +196,10 @@ llvm::ModulePassManager {{cookiecutter.target_name.capitalize()}}PassMachinery::
 
   addLateBuiltinsPasses(PM, tuner);
 
-  compiler::utils::HandleBarriersOptions HBOpts;
-  HBOpts.IsDebug = options.opt_disable;
-  HBOpts.ForceNoTail = env_var_opts.force_no_tail;
-  PM.addPass(compiler::utils::HandleBarriersPass(HBOpts));
+  compiler::utils::WorkItemLoopsPassOptions WIOpts;
+  WIOpts.IsDebug = options.opt_disable;
+  WIOpts.ForceNoTail = env_var_opts.force_no_tail;
+  PM.addPass(compiler::utils::WorkItemLoopsPass(WIOpts));
 
   compiler::addPrepareWorkGroupSchedulingPasses(PM);