+++ modules/compiler/source/base/source/pass_pipelines.cpp
@@ -56,10 +56,10 @@ void addPreVeczPasses(ModulePassManager &PM,
 
   // We need to use the software implementation of the work-group collective
   // builtins. Because ReplaceWGCPass may introduce barrier calls it needs to be
-  // run before PrepareBarriersPass. When using the Handle Barriers pass, we can
-  // run the Replace WGC pass in Scans Only mode, since the Handle Barriers pass
+  // run before PrepareBarriersPass. When using the WorkItemLoopsPass, we can
+  // run the Replace WGC pass in Scans Only mode, since the WorkItemLoopsPass
   // has its own implementations of reductions and broadcasts.
-  PM.addPass(compiler::utils::ReplaceWGCPass(tuner.handling_barriers));
+  PM.addPass(compiler::utils::ReplaceWGCPass(tuner.handling_work_item_loops));
 
   // We have to inline all functions containing barriers before running vecz,
   // because the barriers in both the scalar and vector kernels need to be